% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Algae.R
\docType{class}
\name{Algae-class}
\alias{Algae-class}
\title{Generic Algae class}
\description{
Only used for technical purposes
}
\seealso{
\linkS4class{AlgaeWeberScenario}, \linkS4class{AlgaeTKTDScenario}
}
