% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Algae.R
\name{Algae_TKTD}
\alias{Algae_TKTD}
\title{Algae model with exponential growth, forcings (P, I) and scaled damage}
\usage{
Algae_TKTD()
}
\description{
Constructor to ease creation of \emph{Algae} ‘EffectScenario’s. Model is a variant
of the \code{Algae_Weber} model (Weber 2012) as cited in EFSA TKTD opinion (2018).
This Algae model, \code{Algae_TKTD()}, a) provides an additional
possibility to simulate the dose-response curve (probit) and b) considers an
scaled internal damage instead of the external concentration.
}
\section{State variables}{

The model has four state variables:
\itemize{
\item \code{A}, Biomass (µg fresh wt)
\item \code{Q}, Mass of phosphorous internal (mg P/μg fresh wt)
\item \code{P}, Mass of phosphorous external (mg P/L)
\item \code{Dw}, Damage concentration (ug/L)
}
}

\section{Model parameters}{

\itemize{
\item Growth model
\itemize{
\item \code{mu_max}, Maximum growth rate (d-1)
\item \code{Q_min}, Minimum intracellular P (mg P/μg fresh wt)
\item \code{Q_max}, Maximum intracellular P (mg P/μg fresh wt)
\item \code{v_max}, Maximum P-uptake rate at non-limited growth (mg P/μg fresh wt/d)
\item \code{k_s},   Half-saturation constant for extracellular P (mg P/L)
\item \code{m_max}, Natural mortality rate (1/d)
\item \code{I_opt}, Optimum light intensity for growth (μE/m²/s)
\item \code{T_opt}, Optimum temperature for growth (°C)
\item \code{T_max}, Maximum temperature for growth (°C)
\item \code{T_min}, Minimum temperature for growth (°C)
}
\item Concentration response (Toxicodynamics)
\itemize{
\item \code{EC_50}, Effect concentration of 50\% inhibition of growth rate (ug L-1)
\item \code{b}, slope of concentration effect curve at EC_50 (-)
\item \code{dose_resp}, shape of the dose response curve (0 = logit, 1 = probit)
}
\item External concentration (Toxicokinetics)
\itemize{
\item \code{kD}, dominant rate constant (used in \code{Algae_TKTD()}only) (d-1)
}
}
}

\section{Forcings}{

Besides exposure events (Cw), the \emph{Algae} model requires two environmental
properties as time-series input: Irradiance (\code{I}, μE/m²/s) and
temperature (\code{T_act}, deg C).
Forcings time-series are represented by \code{data.frame} objects
consisting of two columns. The first for time and the second for the
environmental factor in question. The input format for all forcings is a
list of the data frames.
}

\section{Simulation output}{


Simulation results will contain state variables.
}

\references{
Weber D, Schaeffer D, Dorgerloh M, Bruns E, Goerlitz G, Hammel K, Preuss TG
and Ratte HT, 2012. Combination of a higher-tier flow-through system and
population modeling to assess the effects of time-variable exposure of
isoproturon on the green algae Desmodesmus subspictatus and
Pseudokirchneriella subcapitata. Environmental Toxicology and
Chemistry, 31, 899–908.

\linkS4class{Transferable},
\linkS4class{EffectScenario}
}
