% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Algae.R
\docType{class}
\name{AlgaeSimpleScenario-class}
\alias{AlgaeSimpleScenario-class}
\title{Algae model with exponential growth without forcings}
\description{
Please refer to \code{\link[=Algae_Simple]{Algae_Simple()}} for details about the model. Class slots are
documented in \linkS4class{EffectScenario}.
}
