% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Algae.R
\docType{class}
\name{AlgaeTKTDScenario-class}
\alias{AlgaeTKTDScenario-class}
\title{Algae model with exponential growth and scaled internal damage}
\description{
Please refer to \code{\link[=Algae_TKTD]{Algae_TKTD()}} for details about the model. Class slots are
documented in \linkS4class{EffectScenario}.
}
