% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Algae.R
\docType{class}
\name{AlgaeWeberScenario-class}
\alias{AlgaeWeberScenario-class}
\title{Algae model with exponential growth and forcings (P, I)}
\description{
Please refer to \code{\link[=Algae_Weber]{Algae_Weber()}} for details about the model. Class slots are
documented in \linkS4class{EffectScenario}.
}
