% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{americamysis}
\alias{americamysis}
\title{A DEB abj scenario of Americamysis bahia}
\format{
An object of class \code{DebAbj} of length 1.
}
\source{
\url{https://www.bio.vu.nl/thb/deb/deblab/add_my_pet/entries_web/Americamysis_bahia/Americamysis_bahia_res.html}
}
\usage{
americamysis
}
\description{
Species parameters were collected from the AddMyPet database entry on
Americamysis bahia (Opossum shrimp). The exposure series consists of a
constant exposure resulting in medium effects on length and reproduction.
}
\seealso{
\code{\link[=DEB_abj]{DEB_abj()}}
}
\keyword{datasets}
