% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_controls.R
\name{cache_controls}
\alias{cache_controls}
\title{Cache control simulations}
\usage{
cache_controls(x, windows, skipZeroExposure = FALSE, ...)
}
\arguments{
\item{x}{vector of \link{scenario} objects}

\item{windows}{\code{list} of window tuples}

\item{skipZeroExposure}{\code{logical}, if \code{TRUE}, windows with zero exposure will
not be included in calculations}

\item{...}{additional parameters passed on to \code{\link[=effect]{effect()}}}
}
\value{
Modified \link{scenario} objects
}
\description{
Cache control simulations
}
