% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CalibrationSet.R
\docType{class}
\name{CalibrationSet-class}
\alias{CalibrationSet-class}
\title{CalibrationSet class}
\description{
CalibrationSet class
}
\section{Slots}{

\describe{
\item{\code{scenario}}{any kind of \code{EffectScenario} or \code{\link[=sequence]{sequence()}}}

\item{\code{data}}{\code{data.frame} with two columns, 1st column with time points and
2nd column with numerical data to fit to}

\item{\code{weight}}{\code{numeric} weight to be applied to this datasets' error term
during fitting}
}}

