% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth.R
\name{ci_from_hessian}
\alias{ci_from_hessian}
\title{Parameter confidence intervals from fit}
\usage{
ci_from_hessian(fit, dof, level = 0.95)
}
\arguments{
\item{fit}{return value from \code{\link[=calibrate]{calibrate()}}}

\item{dof}{integer, Degrees of Freedom, commonly the number of independent observations
minus the number of fitted parameters}

\item{level}{numeric, desired confidence level, i.e a value between zero and one.
Defaults to 95\% (\code{0.95}).}
}
\value{
name list of numeric vectors of length two, the elements
representing the lower and upper confidence limit, respectively.
}
\description{
Calculates parameter confidence intervals for a fit returned by \code{\link[=calibrate]{calibrate()}}.
The fit must provide a valid \emph{Hessian} matrix.
}
