% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man-deb.R
\name{DEB-models}
\alias{DEB-models}
\alias{Deb-class}
\title{Dynamic Energy Budget (DEB) models}
\description{
Supported models:
\itemize{
\item \link{DEB_abj}
\item \link{DEBtox}
}
}
\seealso{
Other DEB models: 
\code{\link{DEB_abj}()},
\code{\link{DEBtox}()}

Other models: 
\code{\link{Algae-models}},
\code{\link{GUTS-RED-models}},
\code{\link{Lemna-models}},
\code{\link{Macrophyte-models}},
\code{\link{Myriophyllum-models}}
}
\concept{DEB models}
\concept{models}
