% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{derive_range}
\alias{derive_range}
\title{Simulates parameter sample and derives uncertainty range}
\usage{
derive_range(model_base, param_sample, obs_full = NULL, ...)
}
\arguments{
\item{model_base}{effect scenario object with mean parameters}

\item{param_sample}{data.frame with parameter sample}

\item{obs_full}{\code{data.frame}, full experimental data set including replicates}

\item{...}{any additional parameters}
}
\value{
data frame with confidence intervals
}
\description{
Simulates parameter sample and derives uncertainty range
}
\examples{
\dontrun{
derive_range(model_base,
               param_sample,
               obs_full,
               times = sim_times
               )
}
}
