% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{diagnostics}
\alias{diagnostics}
\alias{diagnostics.default}
\alias{diagnostics.cvasi_simulate}
\title{Diagnostics of solvers}
\usage{
diagnostics(obj, ...)

\method{diagnostics}{default}(obj, ...)

\method{diagnostics}{cvasi_simulate}(obj, ...)
}
\arguments{
\item{obj}{return value of a simulation}

\item{...}{unused parameters}
}
\description{
Prints several diagnostics of the simulation to the console, e.g. number of
steps taken, the last step size, etc. The information is provided by
\code{\link[deSolve:diagnostics]{deSolve::diagnostics()}}.
}
\seealso{
\code{\link[deSolve:diagnostics]{deSolve::diagnostics()}}
}
