% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmagna}
\alias{dmagna}
\title{A DEBtox scenario of Daphnia magna}
\format{
An object of class \code{DebTox} of length 1.
}
\usage{
dmagna
}
\description{
Species and substance parameters were collected from test runs of the
original \link{DEBtox} Daphnia model.
}
\seealso{
\code{\link[=DEBtox]{DEBtox()}}
}
\keyword{datasets}
