% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effect}
\alias{effect}
\title{Effect level}
\usage{
effect(x, factor = 1, max_only = TRUE, ep_only = FALSE, marginal_effect, ...)
}
\arguments{
\item{x}{vector of \code{EffectScenario} objects}

\item{factor}{optional numeric value which scales the exposure time-series}

\item{max_only}{\code{logical}, if \code{TRUE} only the maximum effect is returned, else
results for all effect windows are reported}

\item{ep_only}{logical, if TRUE only effect endpoints are returned as a vector}

\item{marginal_effect}{\code{numeric}, if set, any effect smaller than this threshold will
be reported as zero to exclude pseudo-effects originating from small numerical
errors}

\item{...}{additional parameters passed on to \code{\link[=simulate]{simulate()}}}
}
\value{
a \code{tibble}, by default containing scenarios, effect levels, and the
exposure window where the maximum effect level occurred. The number of columns
depends on the enabled effect endpoints and function arguments.

By default, the first column, named \code{scenarios}, contains the original scenario
objects that were the basis of the calculation. For each effect endpoint, it
will be followed by one column with the maximum effect level and two columns
containing start and end time of the associated exposure window. If exposure
windows are disabled, the columns will just contain the start and end time of
the simulation. The effect level column will have the name of the effect
endpoint, start and end time will additionally have the suffixes \code{.dat.start}
and \code{.dat.end}, respectively.
}
\description{
Derives the effect level due to toxicant exposure in the supplied scenarios.
Either relative to a control scenario or derived directly from model
endpoints, depending on model type. For scenarios with moving exposure windows,
the maximum effect is returned.
}
\details{
By default, only the maximum effect in all moving exposure windows will
be returned. If argument \code{max_only=FALSE} is set, the returned table will
be converted to long-format and will contain effect levels for each
assessed exposure window.
\subsection{Output formatting}{

Start and end time of exposure windows can be disabled by setting \code{ep_only=TRUE}.
Effect levels smaller than a certain threshold can be automatically set to
zero (\code{0.0}) to avoid spurious effect levels introduced by numerical errors.
Set \code{marginal_effect} to an adequate value less than 1\%.
}

\subsection{Computational efficiency}{

Calculations can be sped up by providing a \code{data.frame} of pre-calculated
control scenarios for each assessed time window. As control scenarios are
by definition independent of any exposure multiplication factor, they can
be reused for repeated calculations, e.g. to derive effect profiles or
dose-response relationships.
}
}
