% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epx.R
\name{epx_min_win}
\alias{epx_min_win}
\title{Detect worst case EPx window}
\usage{
epx_min_win(epx_mtw, only_first = TRUE)
}
\arguments{
\item{epx_mtw}{result of \code{epx_mtw()}}

\item{only_first}{TRUE = first window only}
}
\value{
a tibble with
\itemize{
\item window.start
\item window.end
\item endpoint
\item level
\item EPx
}
}
\description{
Filters the minimal EPx per endpoint and level from the \code{epx_mtw} object and
returns it as the worst case windows. If there is more than one worst case
window, only the first window will be returned (if \code{only_first == TRUE}), or
all windows will be returned (if \code{only_first == FALSE}).
}
\examples{
\dontrun{
ti <- 0:21
expo <- abs(0.01*ti + rnorm(length(ti), 0, 0.05))
exposure <- data.frame(time = ti, conc = expo)
metsulfuron_epx_mtw <- metsulfuron \%>\%
  set_exposure(exposure) \%>\%
  epx_mtw(level = 10, factor_cutoff = 1000,
    window_length = 7, window_interval = 1)

metsulfuron_epx_mtw

metsulfuron_epx_mtw \%>\% epx_min_win()
}
}
