% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ExposureSeries.R
\docType{class}
\name{ExposureSeries-class}
\alias{ExposureSeries-class}
\alias{ExposureSeries}
\title{Exposure time-series}
\usage{
ExposureSeries(series, dates, file, meta, context)
}
\arguments{
\item{series}{\code{data.frame} with two columns containing a time-series}

\item{dates}{\code{vector}, optional original list of time stamps}

\item{file}{\code{character}, optional file name where data originates from}

\item{meta}{\code{list}, optional metadata}

\item{context}{\code{list} optional contextual metadata such as project ids}
}
\description{
The \code{ExposureSeries} class encapsulates an exposure time-series with its
metadata, such as formatted datetime strings and file name where the
series was loaded from. The constructor \code{ExposureSeries()} can be used to
ease object creation. \code{no_exposure()} is shorthand to create a time-series
of constant zero exposure.

Eases the creation of \linkS4class{ExposureSeries} objects. Can be used
to initialize the object's slots.
}
\section{Slots}{

\describe{
\item{\code{dates}}{original time points of time-series, e.g. time stamps of the form \verb{2000-01-01 12:00}}

\item{\code{file}}{\code{character}, file name where data originates from, may be empty}

\item{\code{meta}}{\code{list}, contains metadata}

\item{\code{context}}{\code{list}, contains contextual metadata, such as project ids}

\item{\code{series}}{\code{data.frame} containing the actual time-series}
}}

