% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_model}
\alias{get_model}
\alias{get_model,ANY-method}
\alias{get_model,list-method}
\alias{get_model,EffectScenario-method}
\alias{get_model,ParameterSet-method}
\alias{get_model_name}
\title{Get model name}
\usage{
get_model(x)
}
\arguments{
\item{x}{(vector of) \link{scenarios} or \code{\link{parameter_set}} objects}
}
\value{
vector of \code{character}
}
\description{
Returns the unique model name that is associated with a scenario, e.g.
\code{GUTS-RED-IT}. The function supports vectorized arguments.
}
\examples{
# returns `GUTS-RED-IT`
get_model(minnow_it)
}
