% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_model_name}
\alias{get_model_name}
\alias{get_model_name,ANY-method}
\alias{get_model_name,list-method}
\alias{get_model_name,EffectScenario-method}
\alias{get_model_name,parameter_set-method}
\alias{get_model}
\title{Returns a model name}
\usage{
get_model_name(x)

\S4method{get_model_name}{ANY}(x)

\S4method{get_model_name}{list}(x)

\S4method{get_model_name}{EffectScenario}(x)

\S4method{get_model_name}{parameter_set}(x)

get_model_name(x)
}
\arguments{
\item{x}{\linkS4class{EffectScenario} or \code{\link{parameter_set}} objects}
}
\value{
\code{character} code identifying a model, e.g. \code{GUTS-RED-IT}
}
\description{
The function supports vectorized inputs. It will raise an error if
non-supported objects are passed as arguments.
}
\examples{
# returns `GUTS-RED-IT`
get_model_name(minnow_it)
}
