% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_tag}
\alias{get_tag}
\alias{get_tag,ANY-method}
\alias{get_tag,list-method}
\alias{get_tag,EffectScenario-method}
\alias{get_tag,parameter_set-method}
\title{Returns a scenario tag}
\usage{
get_tag(x)

\S4method{get_tag}{ANY}(x)

\S4method{get_tag}{list}(x)

\S4method{get_tag}{EffectScenario}(x)

\S4method{get_tag}{parameter_set}(x)
}
\arguments{
\item{x}{\linkS4class{EffectScenario} or \code{\link{parameter_set}} objects}
}
\value{
\code{character} custom tag identifying a scenario
}
\description{
The function supports vectorized inputs. It will raise an error if
non-supported objects are passed as arguments.
}
\examples{
# returns `fathead minnow`
get_tag(minnow_it)
}
