% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_times.R
\name{get_times}
\alias{get_times}
\alias{get_times,ANY-method}
\alias{get_times,list-method}
\alias{get_times,EffectScenario-method}
\alias{get_times,ScenarioSequence-method}
\title{Get output times}
\usage{
get_times(x)
}
\arguments{
\item{x}{(vector of) \link{scenario} objects}
}
\value{
(list of) times vector
}
\description{
Get output times
}
\examples{
# Create a scenario
myscenario <- GUTS_RED_IT() \%>\% set_times(0:5)
# Returns the defined output times
get_times(myscenario)
}
\seealso{
\code{\link[=set_times]{set_times()}}
}
