% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_exposure_text}
\alias{import_exposure_text}
\title{Import exposure profiles from text files}
\usage{
import_exposure_text(pathtofiles, profileNames = NA, sep = "\\t")
}
\arguments{
\item{pathtofiles}{path to files}

\item{profileNames}{the names of the profiles}

\item{sep}{the field separator character}
}
\value{
list of exposure profiles
}
\description{
Read several exposure profiles at their corresponding paths at 'pathtofiles'.
The exposure profiles are read for each file that are character seperated
text files in a table format. The expected table format includes the columns
\code{time} and \code{conc} for the concentration at each timestep.
}
