% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_toxswa.R
\name{import_toxswa}
\alias{import_toxswa}
\title{Read TOXSWA exposure profiles from out-files}
\usage{
import_toxswa(pathtofiles, profileNames = NA)
}
\arguments{
\item{pathtofiles}{paths to toxswa files that should be read}

\item{profileNames}{a vector of names to be used for the loaded toxswa files. if "NA" the filename without extension will be used}
}
\value{
list of exposure profiles each with two columns (time and concentration)
}
\description{
Read several TOXSWA exposure profiles from out-files at their corresponding
paths given by'pathtofiles'.
}
