% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_deb}
\alias{is_deb}
\alias{is_DEB}
\title{Test if argument is a DEB scenario}
\usage{
is_deb(x)

is_DEB(x)
}
\arguments{
\item{x}{vector of \code{EffectScenario} objects}
}
\value{
vector of \code{logical} values
}
\description{
Test if argument is a DEB scenario
}
\section{Functions}{
\itemize{
\item \code{is_DEB()}: Deprecated alias.

}}
