% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_guts}
\alias{is_guts}
\alias{is_GUTS}
\alias{is_guts_it}
\alias{is_GUTS_IT}
\alias{is_guts_sd}
\alias{is_GUTS_SD}
\title{Test if argument is a GUTS scenario}
\usage{
is_guts(x)

is_GUTS(x)

is_guts_it(x)

is_GUTS_IT(x)

is_guts_sd(x)

is_GUTS_SD(x)
}
\arguments{
\item{x}{vector of \code{EffectScenario} objects}
}
\value{
vector of \code{logical} values
}
\description{
Test if argument is a GUTS scenario
}
\section{Functions}{
\itemize{
\item \code{is_GUTS()}: Deprecated alias.

\item \code{is_guts_it()}: Test if argument is a GUTS-IT scenario

\item \code{is_GUTS_IT()}: Deprecated alias.

\item \code{is_guts_sd()}: Test if argument is a GUTS-SD scenario

\item \code{is_GUTS_SD()}: Deprecated alias.

}}
\examples{
# returns `TRUE`
is_guts(minnow_it)
is_guts(GUTS_RED_IT())

# returns `c(TRUE,TRUE,TRUE)`
is_guts(c(minnow_it, minnow_it, minnow_it))

# returns `FALSE`
is_guts_sd(minnow_it)
}
