% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_GUTS_IT}
\alias{is_GUTS_IT}
\title{Test if scenario/model is an Internal Threshold (IT) type}
\usage{
is_GUTS_IT(x)
}
\arguments{
\item{x}{vector of \code{EffectScenario} objects}
}
\value{
vector of logical values
}
\description{
Test if scenario/model is an Internal Threshold (IT) type
}
\examples{
is_GUTS_IT(GUTS_RED_IT())
# returns TRUE
}
