% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_GUTS_SD}
\alias{is_GUTS_SD}
\title{Test if scenario/model is a Stochastic Death (SD) type}
\usage{
is_GUTS_SD(x)
}
\arguments{
\item{x}{vector of \code{EffectScenario} objects}
}
\value{
vector of logical values
}
\description{
Test if scenario/model is a Stochastic Death (SD) type
}
\examples{
is_GUTS_SD(GUTS_RED_SD())
# returns TRUE
}
