% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_lemna}
\alias{is_lemna}
\alias{is_Lemna}
\title{Test if argument is a Lemna scenario}
\usage{
is_lemna(x)

is_Lemna(x)
}
\arguments{
\item{x}{vector of \link{scenario} objects}
}
\value{
vector of logical values
}
\description{
Test if argument is a Lemna scenario
}
\section{Functions}{
\itemize{
\item \code{is_Lemna()}: Deprecated alias.

}}
