% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_LemnaThreshold}
\alias{is_LemnaThreshold}
\title{Test if argument is a LemnaThreshold scenario}
\usage{
is_LemnaThreshold(x)
}
\arguments{
\item{x}{vector of \link{scenarios} objects}
}
\value{
vector of \code{logical} values
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\seealso{
\code{\link[=is_lemna]{is_lemna()}}
}
