% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_scenario}
\alias{is_scenario}
\title{Test if argument is an effect scenario}
\usage{
is_scenario(x)
}
\arguments{
\item{x}{Some value or object}
}
\value{
vector of \code{logical} values
}
\description{
Supports vectorized arguments.
}
\examples{
# returns `TRUE`
is_scenario(minnow_it)

# returns `FALSE`
is_scenario(list())

}
