% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Lemna.R
\docType{class}
\name{Lemna-class}
\alias{Lemna-class}
\title{Generic Lemna class}
\description{
Only used for technical purposes
}
\seealso{
\linkS4class{LemnaSchmittScenario}, \linkS4class{LemnaSetacScenario}
}
