% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Lemna.R
\docType{class}
\name{LemnaSchmittScenario-class}
\alias{LemnaSchmittScenario-class}
\title{Lemna model class (Schmitt et al. 2013)}
\description{
Please refer to \code{\link[=Lemna_Schmitt]{Lemna_Schmitt()}} for details about the model. Class slots are
documented in \linkS4class{EffectScenario}.
}
