% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Lemna.R
\docType{class}
\name{LemnaSetacScenario-class}
\alias{LemnaSetacScenario-class}
\title{Lemna model class (Klein et al. 2021)}
\description{
Please refer to \code{\link[=Lemna_SETAC]{Lemna_SETAC()}} for details about the model. Class slots are
documented in \linkS4class{EffectScenario}.
}
