% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_enable}
\alias{log_enable}
\alias{log_disable}
\title{Start and stop logging}
\usage{
log_enable(file = NULL, append = TRUE, envir = parent.frame())

log_disable()
}
\arguments{
\item{file}{\code{character}, file name or path to a log file}

\item{append}{\code{logical}, if \code{TRUE} output will be appended to an existing log file,
otherwise the log file will be replaced}

\item{envir}{log will be automatically disabled if \code{environment} is exited,
set to \code{NULL} to disable}
}
\value{
no return value
}
\description{
Start and stop logging
}
