% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_msg}
\alias{log_msg}
\title{Add a log message}
\usage{
log_msg(...)
}
\arguments{
\item{...}{elements will be concatenated using \code{paste0()}}
}
\value{
no return value
}
\description{
Message will only appear in the console or in log file if logging was
enabled using \code{log_enable()}.
}
\examples{
log_msg("this message will not appear")

log_enable()
log_msg("this message will appear")
log_msg("a number of ","elements to ",42," concatenate")
}
