% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_scenarios}
\alias{log_scenarios}
\title{Log scenario properties}
\usage{
log_scenarios(x, header = TRUE)
}
\arguments{
\item{x}{vector of \code{EffectScenario} objects}

\item{header}{\code{logical}, if \code{TRUE} a header line will be printed}
}
\value{
unmodified argument \code{x}
}
\description{
Log scenario properties
}
