% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-magma.R
\docType{class}
\name{Magma-class}
\alias{Magma-class}
\title{Magma scenario class}
\description{
This entry documents the class definition used for \link{Magma}-type scenarios.
For details regarding the model, please refer to the \link[=Magma]{Magma model} manual.
}
\section{Slots}{

\describe{
\item{\code{growth_model}}{\emph{character}, selects the growth model, such as \emph{exponential}
or \emph{logistic} growth.}
}}

\seealso{
\link{Macrophyte-models}
}
