% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metsulfuron}
\alias{metsulfuron}
\title{Lemna data published by Schmitt (2013)}
\format{
An object of class \code{LemnaSchmittScenario} of length 1.
}
\usage{
metsulfuron
}
\description{
Data set for the parametrisation of a mechanistic combined
toxicokinetic-toxicodynamic (TK/TD) and growth model for the aquatic
macrophytes Lemna spp. as published by Schmitt \emph{et al.} (2013).
The growth model was parameterised by Schmitt et al. based on these
data while toxicokinetic and toxicodynamic parameters were determined by
calibrating the model using substance specific effect data of metsulfuron-methyl.
}
\references{
Schmitt W., Bruns E., Dollinger M., and Sowig P., 2013: \emph{Mechanistic TK/TD-model
simulating the effect of growth inhibitors on Lemna populations}. Ecol Model 255,
pp. 1-10. \doi{10.1016/j.ecolmodel.2013.01.017}
}
\seealso{
\link{Lemna-models}
}
\keyword{datasets}
