% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Myriophyllum.R
\docType{class}
\name{MyrioExpScenario-class}
\alias{MyrioExpScenario-class}
\title{Myriophyllum model class}
\description{
Please refer to \code{\link[=Myrio]{Myrio()}} for details about the model. Class slots are
documented in \linkS4class{EffectScenario}.
}
