% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Myriophyllum.R
\docType{class}
\name{MyrioLogScenario-class}
\alias{MyrioLogScenario-class}
\title{Myriophyllum model class}
\description{
Please refer to \code{\link[=Myrio_log]{Myrio_log()}} for details about the model. Class slots are
documented in \linkS4class{EffectScenario}.
}
