% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Myriophyllum.R
\docType{class}
\name{Myriophyllum-class}
\alias{Myriophyllum-class}
\title{Myriophyllum model class}
\description{
Only used for technical purposes
}
\seealso{
\linkS4class{MyrioExpScenario}, \linkS4class{MyrioLogScenario}
}
