% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-myriophyllum.R
\name{Myriophyllum-models}
\alias{Myriophyllum-models}
\alias{Myriophyllum-class}
\title{Myriophyllum models}
\description{
Supported models:
\itemize{
\item \code{\link[=Myrio]{Myrio()}}, with exponential growth
\item \code{\link[=Myrio_log]{Myrio_log()}}, with logistic growth
}
}
\seealso{
\link{Lemna-models}, \link{Transferable}

Other Myriophyllum models: 
\code{\link{Myrio}()},
\code{\link{Myrio_log}()}

Other models: 
\code{\link{Algae-models}},
\code{\link{DEB-models}},
\code{\link{GUTS-RED-models}},
\code{\link{Lemna-models}},
\code{\link{Macrophyte-models}}
}
\concept{Myriophyllum models}
\concept{models}
