% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-parameter_set.R
\docType{class}
\name{parameter_set-class}
\alias{parameter_set-class}
\title{Set of model parameters}
\description{
Set of model parameters
}
\section{Slots}{

\describe{
\item{\code{model}}{\code{character}, a string containing a model name, e.g. \code{"GUTS-RED-IT"}}

\item{\code{tag}}{\code{character}, an optional identifier}

\item{\code{param}}{named \code{list} of model parameters}
}}

\examples{
parameter_set("GUTS-RED-IT", list(kd=0.01, hb=0))
}
