% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.param_space}
\alias{plot.param_space}
\alias{plot_param_space}
\title{Plot profiled parameter space}
\usage{
\method{plot}{param_space}(x, y, ...)

plot_param_space(x)
}
\arguments{
\item{x}{Return value of \code{\link[=explore_space]{explore_space()}}}

\item{y}{Unused argument}

\item{...}{Unused arguments}
}
\value{
plot object
}
\description{
The function provides bivariate parameter space plots indicating
parameter draws (from the 95\% confidence intervals per parameter obtained
through likelihood profiling) that fall within the inner rim (in green, i.e.
parameter sets which are not significantly different from the original, based
on a chi-square test). The original parameter set is also indicated (in orange),
and, if different from the original set, the best fit parameter set is indicated
(in red).
}
\section{Functions}{
\itemize{
\item \code{plot_param_space()}: Alias of \code{plot.param_space()} for backwards-compatibility.

}}
