% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_profile.R
\name{plot_lik_profile}
\alias{plot_lik_profile}
\title{Plot likelihood profiles or all profiled parameters}
\usage{
plot_lik_profile(x)
}
\arguments{
\item{x}{object of class lik_profile}
}
\value{
plots
}
\description{
The function provides a combined plot of the likelihood profiles
of all parameters profiled.
}
