% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R
\name{plot_param_space}
\alias{plot_param_space}
\title{Plot likelihood profiles or all profiled parameters}
\usage{
plot_param_space(x)
}
\arguments{
\item{x}{object of class param_space}
}
\value{
plots
}
\description{
The function provides bivariate parameter space plots indicating
parameter draws (from the 95\%confidence intervals per parameter obtained
through likelihood profiling) that fall within the inner rim (in green, i.e.
parameter sets which are not significantly different from the original, based
on a chi-square test). The original parameter set is also indicated (in orange),
and, if different from the original set, the best fit parameter set is indicated
(in red)
}
