% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_ppc_combi}
\alias{plot_ppc_combi}
\title{Create PPC plot for one or more datasets}
\usage{
plot_ppc_combi(table, xy_lim = NULL)
}
\arguments{
\item{table}{\code{data.frame} containing return values of calls to \code{plot_ppc()}}

\item{xy_lim}{optional \code{numeric}, limits of x and y axis for plotting}
}
\description{
The function expects a data.frame with four mandatory and one optional
column. The mandatory columns are as follows:
\itemize{
\item \code{pred}: mean of predictions e.g. frond number for lemna
\item \code{max}: maximum of predictions
\item \code{min}: minimum of predictions
\item \code{obs}: observations
The optional column is to be named \code{study} and contains a study identifier.
If more than one study identifier is present in the table, individual
studies will be plotted in different colors and a legend will be displayed.
}
}
