% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_scenario}
\alias{plot_scenario}
\title{Creates a prediction plot for one effect scenario}
\usage{
plot_scenario(model_base, plot_col = 2, trial_number = NULL)
}
\arguments{
\item{model_base}{effect scenario object with mean parameters}

\item{plot_col}{output column which should be plotted, default = 2}

\item{trial_number}{name for model run (if available tag is used)}
}
\value{
plot of the results for one \verb{effect scenario}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Sometimes it is helpful if the user can plot results of one effect
scenario. This is for instance the case for test simulations or predictions
for one profile. This function runs the simulation for one effect scenario
and plots the results. Function plots the time (column 1) and the predictions
(column 2, can be changed by the user plot_col)
}
\examples{
plot_scenario(metsulfuron)
}
