% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull_metadata}
\alias{pull_metadata}
\title{Pull metadata from scenarios}
\usage{
pull_metadata(x, model = TRUE, exposure = TRUE)
}
\arguments{
\item{x}{an \linkS4class{EffectScenario} vector or a \code{data.frame} containing a column
\code{scenario} with \code{EffectScenario} objects}

\item{model}{\code{logical}, if \code{TRUE} then model metadata is pulled}

\item{exposure}{\code{logical}, if \code{TRUE} then exposure series metadata is pulled}
}
\value{
a \code{data.frame}
}
\description{
The method pulls available metadata from scenario objects and returns a
table with additional columns. If the argument already was a \code{data.frame}
object, the columns are appended. May overwrite existing columns of the
same name.
}
\examples{
metsulfuron \%>\%
  pull_metadata()
}
