% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ringtest_sd}
\alias{ringtest_sd}
\title{EFSA Ringtest Dataset A SD}
\format{
An object of class \code{data.frame} with 42 rows and 4 columns.
}
\usage{
ringtest_sd
}
\description{
This dataset is part of EFSA's \emph{GUTS} ringtest to compare results from software
implementations of \link[=GUTS-RED-models]{GUTS-RED models} (EFSA 2018). The ringtest
focused on \link[=GUTS_RED_IT]{GUTS-RED-IT} and \link[=GUTS_RED_SD]{GUTS-RED-SD} models.
}
\details{
\emph{Dataset A SD} for the GUTS ring test is a Hypothetical data set constructed with SD.
}
\references{
EFSA PPR Panel (EFSA Panel on Plant Protection Products and their Residues),
Ockleford C, Adriaanse P, Berny P, et al., 2018: \emph{Scientific Opinion on the
state of the art of Toxicokinetic/Toxicodynamic (TKTD) effect models for
regulatory risk assessment of pesticides for aquatic organisms}. EFSA Journal 2018;
16(8):5377, 188 pp. \doi{10.2903/j.efsa.2018.5377}

https://cran.r-project.org/web/packages/GUTS/vignettes/ringTest.html
}
\seealso{
\link{GUTS-RED-models}
}
\keyword{datasets}
