% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rsubcapitata}
\alias{rsubcapitata}
\title{A Weber scenario of algae exposed to isoproturon}
\format{
An object of class \code{AlgaeWeber} of length 1.
}
\usage{
rsubcapitata
}
\description{
The scenario recreates the conditions of algae species \emph{R. subcapitata} exposed
to isoproturon in a flow-through reactor experiment, see Weber et al. (2012)
publication. Scenario parameters and exposure series were set according to
reported values to represent conditions in reactor A, as depicted in
Figure 4A in Weber et al.
}
\references{
Weber D, Schaefer D, Dorgerloh M, Bruns E, Goerlitz G, Hammel K, Preuss TG
and Ratte HT, 2012. Combination of a higher-tier flow-through system and
population modeling to assess the effects of time-variable exposure of
isoproturon on the green algae Desmodesmus subspictatus and
Pseudokirchneriella subcapitata. Environmental Toxicology and
Chemistry, 31, 899-908. \doi{10.1002/etc.1765}
}
\seealso{
\link{Algae_Weber}
}
\keyword{datasets}
