% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ScenarioSequence.R
\docType{class}
\name{ScenarioSequence-class}
\alias{ScenarioSequence-class}
\title{Class: scenario sequence}
\description{
Represents a sequence of time-continuous scenarios. For more information, please
refer to \code{\link[=sequence]{sequence()}}.
}
\section{Slots}{

\describe{
\item{\code{scenarios}}{list of \linkS4class{EffectScenario} objects}
}}

\examples{
# create two scenarios that need to be simulated one after the other
scen1 <- minnow_it \%>\% set_times(0:3)
scen2 <- minnow_it \%>\% set_times(3:6) \%>\% set_param(c(kd=0))

# create an empty sequence
sq <- sequence()
# assign scenarios to the sequence one by one
scenario(sq, 1) <- scen1
scenario(sq, 2) <- scen2

# simulate the sequence
simulate(sq)

# create a sequence and assign scenarios
sq <- sequence(list(scen1, scen2))
# assign list of scenarios
scenarios(sq) <- list(scen1, scen2)
}
