% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ScenarioSequence.R
\name{scenarios}
\alias{scenarios}
\alias{scenarios<-}
\alias{scenario<-}
\alias{scenarios,ScenarioSequence-method}
\alias{scenarios<-,ScenarioSequence-method}
\alias{scenario<-,ScenarioSequence-method}
\title{Access elements in a scenario sequence}
\usage{
scenarios(x)

scenarios(x) <- value

scenario(x, i) <- value

\S4method{scenarios}{ScenarioSequence}(x)

\S4method{scenarios}{ScenarioSequence}(x) <- value

\S4method{scenario}{ScenarioSequence}(x, i) <- value
}
\arguments{
\item{x}{a \emph{scenario sequence} object}

\item{value}{new scenario(s) to set}

\item{i}{numeric index of an element}
}
\value{
list of scenarios
}
\description{
\code{NULL} values may be used
}
\examples{
# access scenarios individually
sq <- sequence()
scenario(sq, 1) <- minnow_it \%>\% set_times(0:3)
scenario(sq, 2) <- minnow_it \%>\% set_times(3:6)

# retrieve all scenarios in a sequence
scenarios(sq)
# set all scenarios at once
scenarios(sq) <- list(minnow_it, NULL)
}
\seealso{
\code{\link[=sequence]{sequence()}}
}
