% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_exposure.R
\name{set_noexposure}
\alias{set_noexposure}
\title{Set zero exposure}
\usage{
set_noexposure(x)
}
\arguments{
\item{x}{vector of \linkS4class{EffectScenario} objects}
}
\value{
vector of \linkS4class{EffectScenario} objects
}
\description{
The scenarios current exposure is replaced by a constant exposure time-series
of value zero(\code{0.0}). Output times are unaffected.
}
\examples{
# Derive effect size in sample scenario without toxicant exposure
minnow_it \%>\%
  set_noexposure() \%>\%
  effect()
}
