% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_param_bounds.R
\name{set_param_bounds}
\alias{set_param_bounds}
\alias{set_param_bounds,EffectScenario,list-method}
\alias{set_param_bounds,CalibrationSet,list-method}
\alias{set_param_bounds,list,list-method}
\title{Set custom boundaries for model parameters}
\usage{
set_param_bounds(x, pars_bound)

\S4method{set_param_bounds}{EffectScenario,list}(x, pars_bound)

\S4method{set_param_bounds}{CalibrationSet,list}(x, pars_bound)

\S4method{set_param_bounds}{list,list}(x, pars_bound)
}
\arguments{
\item{x}{object(s) to modify}

\item{pars_bound}{named list of lists, where the first level lists the parameters
by name, and the second level lists the lower and upper boundary}
}
\value{
\link{scenario} or \link{CalibrationSet} with modified parameter boundaries
}
\description{
Modifies the boundaries for parameters for one or more \link{scenario} or
\link{CalibrationSet} objects.
}
\examples{
metsulfuron \%>\%
   set_param_bounds(pars_bound = list(k_phot_max = list(0,30),
                                     k_resp = list(0,10)))

}
