% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_tag}
\alias{set_tag}
\title{Set a scenario tag}
\usage{
set_tag(x, tag)
}
\arguments{
\item{x}{\code{vector} of or single \code{EffectScenario} object}

\item{tag}{\code{vector}}
}
\value{
\code{vector} of or single \code{EffectScenario} object
}
\description{
Set a scenario tag
}
\examples{
GUTS_RED_SD() \%>\%
  set_tag("sample tag")
}
