% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Algae.R
\name{Algae-models}
\alias{Algae-models}
\alias{Algae-class}
\title{Algae models}
\description{
Overview of supported \emph{Algae} models
}
\details{
\itemize{
\item \code{\link[=Algae_Weber]{Algae_Weber()}} by Weber \emph{et al.} (2012)
\item \code{\link[=Algae_TKTD]{Algae_TKTD()}} based on Weber \emph{et al.} (2012), but with scaled damage
\item \code{\link[=Algae_Simple]{Algae_Simple()}} Simple growth model without additional forcing variables
}
}
\section{Biomass transfer}{

Models supporting biomass transfer can be instructed to move a fixed amount
of biomass to a new medium after a period of time. This feature replicates
a procedure occurring in e.g. Lemna effect studies and may be necessary to
recreate study results.

The biomass transfer feature assumes that always a fixed amount of
biomass is transferred. Transfers can occur at any fixed point in time or
in regular intervals. During a transfer, the biomass is reset to the
transferred amount and additional compartments can be scaled 1:1 accordingly,
to e.g. reflect the change in internal toxicant mass when biomass is modified.
Transfer settings can be modified using \code{\link[=set_transfer]{set_transfer()}}.

Any transfer time point must also be an output time point. If a transfer
occurs, simulation results of that time point will report the model state
\strong{before} the transfer. Be aware that in order to use transfers at regular
intervals, the simulation must start at time point zero.
}

\references{
Weber D, Schaeffer D, Dorgerloh M, Bruns E, Goerlitz G, Hammel K, Preuss TG
and Ratte HT, 2012. Combination of a higher-tier flow-through system and
population modeling to assess the effects of time-variable exposure of
isoproturon on the green algae Desmodesmus subspictatus and
Pseudokirchneriella subcapitata. Environmental Toxicology and
Chemistry, 31, 899-908. \doi{10.1002/etc.1765}
EFSA Panel on Plant Protection Products and their Residues, 2018. Scientific
opinion on the state of the art of Toxicokinetic/Toxicodynamic (TKTD) effect
models for regulatory risk assessment of pesticides for aquatic organisms.
EFSA journal 16:5377 \doi{10.2903/j.efsa.2018.5377}
}
\seealso{
\link{Lemna-models}, \link{Transferable}

Other algae models: 
\code{\link{Algae_Simple}()},
\code{\link{Algae_TKTD}()},
\code{\link{Algae_Weber}()}

Other scenarios: 
\code{\link{DEB-models}},
\code{\link{GUTS-RED-models}},
\code{\link{Lemna-models}},
\code{\link{Macrophyte-models}},
\code{\link{Myriophyllum-models}},
\code{\link{Scenarios}},
\code{\link{Transferable}}
}
\concept{algae models}
\concept{scenarios}
