% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Algae.R
\name{Algae-models}
\alias{Algae-models}
\title{Algae models}
\description{
Overview of supported \emph{Algae} models
}
\details{
\itemize{
\item \code{\link[=Algae_Weber]{Algae_Weber()}} by Weber \emph{et al.} (2012)
\item \code{\link[=Algae_TKTD]{Algae_TKTD()}} based on Weber \emph{et al.} (2012), but with scaled damage
\item \code{\link[=Algae_Simple]{Algae_Simple()}} Simple growth model without additional forcing variables
}
}
\section{Biomass transfer}{

Models supporting biomass transfer can be instructed to move a fixed amount
of biomass to a new medium after a period of time. This feature replicates
a procedure occurring in e.g. Lemna effect studies and may be necessary to
recreate study results.

The biomass transfer feature assumes that always a fixed amount of
biomass is transferred. Transfers can occur at any fixed point in time or
in regular intervals. During a transfer, the biomass is reset to the
transferred amount and additional compartments can be scaled 1:1 accordingly,
to e.g. reflect the change in internal toxicant mass when biomass is modified.
Transfer settings can be modified using \code{\link[=set_transfer]{set_transfer()}}.

Any transfer time point must also be an output time point. If a transfer
occurs, simulation results of that time point will report the model state
\strong{before} the transfer. Be aware that in order to use transfers at regular
intervals, the simulation must start at time point zero.
}

