% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_controls}
\alias{cache_controls}
\title{Cache control simulations}
\usage{
cache_controls(x, ...)
}
\arguments{
\item{x}{parameter not used}

\item{...}{parameters not used}
}
\value{
nothing
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Handling of cached control simulations has been modified and is solely
managed by package functions. \code{cache_controls()} is no longer needed and
will raise an error if called.
}
