% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Deb.R
\name{DEB-models}
\alias{DEB-models}
\alias{DebScenario-class}
\title{Dynamic Energy Budget (DEB) models}
\description{
Supported models:
\itemize{
\item \link{DEB_abj}
\item \link{DEBtox}
}
}
\seealso{
Other DEB models: 
\code{\link{DEB_abj}()},
\code{\link{DEBtox}()}

Other scenarios: 
\code{\link{Algae-models}},
\code{\link{GUTS-RED-models}},
\code{\link{Lemna-models}},
\code{\link{Macrophyte-models}},
\code{\link{Myriophyllum-models}},
\code{\link{Scenarios}},
\code{\link{Transferable}}
}
\concept{DEB models}
\concept{scenarios}
