% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{diagnostics}
\alias{diagnostics}
\alias{diagnostics.cvasi.simulate}
\title{Print Diagnostic Characteristics of Solvers}
\usage{
\method{diagnostics}{cvasi.simulate}(obj, ...)
}
\arguments{
\item{obj}{is an output data structure produced by one of the
    solver routines.
  }

\item{...}{optional arguments allowing to extend \code{diagnostics}
    as a generic function.
  }
}
\description{
Prints several diagnostics of the simulation to the screen, e.g.
  number of steps taken, the last step size, ...
}
\details{
Detailed information obout the success of a simulation is printed,
  if a \code{diagnostics} function exists for a specific solver routine.
  A warning is printed, if no class-specific diagnostics exists.

  Please consult the class-specific help page for details.
}
\seealso{
\code{\link[deSolve]{diagnostics.deSolve}} for diagnostics of differential
  equaton solvers.
%% enable this when bvpSolve is on CRAN
%  \code{\link[bvpSolve:diagnostics]{diagnostics.bvpSolve}} for
%  diagnostics of boundary value problem solvers.
}
