% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmagna}
\alias{dmagna}
\title{A DEBtox Daphnia magna scenario}
\format{
An object of class \code{DebDaphnia} of length 1.
}
\usage{
dmagna
}
\description{
Species and substance parameters were collected from test runs of the
original DEBtox Daphnia model.
}
\seealso{
\code{\link[=DEB_Daphnia]{DEB_Daphnia()}}
}
\keyword{datasets}
