% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_DEB}
\alias{is_DEB}
\title{Test if scenario/model is a DEB type}
\usage{
is_DEB(x)
}
\arguments{
\item{x}{vector of \code{EffectScenario} objects}
}
\value{
vector of logical values
}
\description{
Test if scenario/model is a DEB type
}
