% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_DEB}
\alias{is_DEB}
\title{Test if argument is a DEB model}
\usage{
is_DEB(x)
}
\arguments{
\item{x}{vector of \code{EffectScenario} objects}
}
\value{
vector of \code{logical} values
}
\description{
Test if argument is a DEB model
}
