% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_GUTS}
\alias{is_GUTS}
\title{Test if scenario/model is a GUTS model}
\usage{
is_GUTS(x)
}
\arguments{
\item{x}{vector of \code{EffectScenario} objects}
}
\value{
vector of logical values
}
\description{
Test if scenario/model is a GUTS model
}
\examples{
is_GUTS(GUTS_RED_IT()) # returns TRUE
is_GUTS(rep(c(GUTS_RED_IT()),3)) # returns c(TRUE,TRUE,TRUE)
}
