% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_Lemna}
\alias{is_Lemna}
\title{Test if scenario/model is a Lemna model}
\usage{
is_Lemna(x)
}
\arguments{
\item{x}{vector of \linkS4class{EffectScenario} objects}
}
\value{
vector of logical values
}
\description{
Also returns \code{TRUE} for \code{LemnaThreshold} models
}
