% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_LemnaThreshold}
\alias{is_LemnaThreshold}
\title{Test if scenario/model is a LemnaThreshold model}
\usage{
is_LemnaThreshold(x)
}
\arguments{
\item{x}{vector of \linkS4class{EffectScenario} objects}
}
\value{
vector of logical values
}
\description{
Test if scenario/model is a LemnaThreshold model
}
