% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_LemnaThreshold}
\alias{is_LemnaThreshold}
\title{Test if argument is a LemnaThreshold model}
\usage{
is_LemnaThreshold(x)
}
\arguments{
\item{x}{vector of \link{scenarios} objects}
}
\value{
vector of \code{logical} values
}
\description{
Test if argument is a LemnaThreshold model
}
\seealso{
\code{\link[=is_Lemna]{is_Lemna()}}
}
