% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minnow_sd}
\alias{minnow_sd}
\title{A GUTS-RED-SD scenario of the fathead minnow}
\format{
An object of class \code{GutsRedSd} of length 1.
}
\source{
\url{https://mosaic.univ-lyon1.fr/guts}
}
\usage{
minnow_sd
}
\description{
The scenario consists of a parameterized GUTS-RED-SD model and a constant
exposure series. Model parameters were derived from an acute fish toxicity
study of the fathead minnow and chlorpyrifos by Geiger \emph{et al.} (1988).
The dataset is also referred to as the \emph{GUTS Ring-test dataset C} and was
downloaded from the \href{https://mosaic.univ-lyon1.fr/guts}{MOSAIC} website.
Fitted parameters were estimated using the \link{morse} package.
}
\details{
The background mortality rate (\code{hb}) was set to zero.
}
\references{
Geiger D.L., Call D.J., and Brooke L.T., 1988: \emph{Acute toxicities of organic
chemicals to fathead minnows (Pimephales promelas): Volume IV}, pp. 195-197.
University of Wisconsin-Superior, Center for Lake Superior Environmental Studies.
ISBN 9780961496838.
}
\seealso{
\link{GUTS-RED-models}
}
\keyword{datasets}
