% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-myriophyllum.R
\name{Myrio}
\alias{Myrio}
\alias{MyrioExp-class}
\alias{MyrioExpScenario-class}
\title{Myriophyllum model with exponential growth}
\usage{
Myrio()
}
\value{
an S4 object of type \linkS4class{MyrioExpScenario}
}
\description{
The \emph{Myriophyllum} model is derived from the \emph{Lemna} TKTD model by
Klein \emph{et al.} (2021). The Myriophyllum model is mathematically equivalent
to the Tier 2C version of the \emph{Lemna} model by Klein \emph{et al.} (2021),
cf. \code{\link[=Lemna_SETAC]{Lemna_SETAC()}}. Recommended settings for Tier 2C are \code{k_photo_fixed=TRUE}
and \code{k_resp=0} (Klein \emph{et al.} 2021).
In particular, the growth model is a simple exponential growth model,
which is considered to be the typical situation for a laboratory macrophyte
study. Instead of frond numbers as for \emph{Lemna}, the biomass is also returned as
total shoot length (\code{TSL}) in simulation results.
Consequently, the model has the additional parameter \code{r_DW_TSL}
(dry weight per total shoot length ratio) instead of \code{r_DW_FN} (dry weight
per frond number ratio).
}
\section{State variables}{

The model has two state variables:
\itemize{
\item \code{BM}, Biomass (g dw m-2 for field studies or mg dw for lab)
\item \code{M_int}, Mass of toxicant in plant population (mass per m2, e.g. ug m-2)
}
}

\section{Model parameters}{

\itemize{
\item Growth model
\itemize{
\item \code{k_photo_max}, Maximum photosynthesis rate (d-1), default: \code{0.47}
}
\item Concentration response (Toxicodynamics)
\itemize{
\item \code{EC50_int}, Internal concentration resulting in 50\% effect (ug L-1)
\item \code{E_max}, Maximum inhibition (-), default: \code{1}
\item \code{b}, Slope parameter (-)
}
\item Internal concentration (Toxicokinetics)
\itemize{
\item \code{P}, Permeability (cm d-1)
\item \code{r_A_DW}, Area per dry-weight ratio (cm2 g-1), default: \code{1000}
\item \code{r_FW_DW}, Fresh weight per dry weight ratio (-), default: \code{16.7}
\item \code{r_FW_V}, Fresh weight density (g cm-3), default: \code{1}
\item \code{r_DW_TSL}, Dry weight per total shoot length ratio  (g (field) or mg (lab) dw cm-1)
\item \code{K_pw}, Partitioning coefficient plant:water (-), default: \code{1}
\item \code{k_met}, Metabolisation rate (d-1), default: \code{0}
}
}
}

\section{Environmental factors}{


None.
}

\section{Parameter boundaries}{

Default values for parameter boundaries are set for all parameters by expert
judgement, for calibration purposes. Values can be modified using \code{\link[=set_bounds]{set_bounds()}}.
}

\section{Simulation output}{

Simulation results will contain the state variables
It is possible to amend the output of \code{\link[=simulate]{simulate()}} with additional model
quantities that are not state variables, for e.g. debugging purposes or to
analyze model behavior. To enable or disable additional outputs, use the
optional argument \code{nout} of \code{\link[=simulate]{simulate()}}. As an example, set \code{nout=2} to
enable reporting of the acceleration factor (\code{MV}) and the mobilization flux
(\code{pC}). Set \code{nout=0} to disable additional outputs (default).

The available output levels are as follows:
\itemize{
\item \code{nout} >= 1: \code{C_int}, internal concentration (mass per volume)
\item \code{nout} >= 2: \code{TSL}, total shoot length (?)
\item \code{nout} >= 3: \code{f_photo}, photosynthesis dependency function (-)
\item Growth and TK/TD
\itemize{
\item \code{nout} >= 4: \code{C_int_unb}, unbound internal concentration (mass per volume)
\item \code{nout} >= 5: \code{C_ext}, external concentration (mass per volume)
}
\item Derivatives
\itemize{
\item \code{nout} >= 6: \code{dBM}, biomass derivative (g dw m-2 d-1)
\item \code{nout} >= 7: \code{dM_int}, mass of toxicant in plants derivative (mass per m2 d-1)
}
}
}

\section{Effects}{

Supported effect endpoints include \emph{BM} (biomass) and \emph{r} (average
growth rate during simulation). The effect on biomass is calculated from
the last state of a simulation. Be aware that endpoint \emph{r} is incompatible
with biomass transfers.
}

\section{Biomass transfer}{

Models supporting biomass transfer can be instructed to move a fixed amount
of biomass to a new medium after a period of time. This feature replicates
a procedure occurring in e.g. Lemna effect studies and may be necessary to
recreate study results.

The biomass transfer feature assumes that always a fixed amount of
biomass is transferred. Transfers can occur at any fixed point in time or
in regular intervals. During a transfer, the biomass is reset to the
transferred amount and additional compartments can be scaled 1:1 accordingly,
to e.g. reflect the change in internal toxicant mass when biomass is modified.
Transfer settings can be modified using \code{\link[=set_transfer]{set_transfer()}}.

Any transfer time point must also be an output time point. If a transfer
occurs, simulation results of that time point will report the model state
\strong{before} the transfer. Be aware that in order to use transfers at regular
intervals, the simulation must start at time point zero.
}

\references{
Klein J., Cedergreen N., Heine S., Reichenberger S., Rendal C.,
Schmitt W., Hommen U., 2021: \emph{Refined description of the Lemna TKTD growth model
based on Schmitt et al. (2013) - equation system and default parameters}.
Report of the working group \emph{Lemna} of the SETAC Europe Interest Group Effect
Modeling. Version 1, uploaded on 22. Sept. 2021.
https://www.setac.org/group/effect-modeling.html
}
\seealso{
\link{Macrophyte-models}, \link{Transferable}, \link{Scenarios}

Other Myriophyllum models: 
\code{\link{Myrio_log}()},
\code{\link{Myriophyllum-models}}

Other macrophyte models: 
\code{\link{Lemna_SETAC}()},
\code{\link{Lemna_Schmitt}()},
\code{\link{Macrophyte-models}},
\code{\link{Myrio_log}()}
}
\concept{Myriophyllum models}
\concept{macrophyte models}
