% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.cvasi.drc}
\alias{plot.cvasi.simulate}
\title{S3 plotting functions}
\usage{
\method{plot}{cvasi.drc}(x, y, scale_x = c("auto", "log10", "none"), ...)

\method{plot}{cvasi.simulate}(x, y, ...)
}
\arguments{
\item{x}{object to plot}

\item{y}{unused parameter}

\item{scale_x}{character, controls how the x-axis is scaled. \code{log10} for
a log10-scaled axis, \code{none} for no scaling, and \code{auto} for automatic selection}

\item{...}{unused parameters}
}
\value{
\emph{ggplot2} plot object
}
\description{
These functions overload \code{\link[base:plot]{base::plot()}} to provide simple plotting
routines to display various time-series and scenario objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{plot(cvasi.drc)}: Plot dose response curves

\item \code{plot(cvasi.simulate)}: Plot return value of \code{\link[=simulate]{simulate()}}

}}
