% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rsubcapitata}
\alias{Rsubcapitata}
\title{An algae scenario}
\format{
An object of class \code{AlgaeTKTDScenario} of length 1.
}
\usage{
Rsubcapitata
}
\description{
Data are from Weber 2012 publication.
}
\references{
Weber D, Schaeffer D, Dorgerloh M, Bruns E, Goerlitz G, Hammel K, Preuss TG
and Ratte HT, 2012. Combination of a higher-tier flow-through system and
population modeling to assess the effects of time-variable exposure of
isoproturon on the green algae Desmodesmus subspictatus and
Pseudokirchneriella subcapitata. Environmental Toxicology and
Chemistry, 31, 899-908. \doi{10.1002/etc.1765}
}
\seealso{
\link{Algae_TKTD}
}
\keyword{datasets}
