% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Schmitt2013}
\alias{Schmitt2013}
\title{A Lemna data set with multiple treatment levels}
\format{
An object of class \code{data.frame} with 56 rows and 4 columns.
}
\usage{
Schmitt2013
}
\description{
Data are from Schmitt 2013 publication.
}
\references{
Schmitt W., Bruns E., Dollinger M., and Sowig P., 2013: \emph{Mechanistic TK/TD-model
simulating the effect of growth inhibitors on Lemna populations}. Ecol Model 255,
pp. 1-10. \doi{10.1016/j.ecolmodel.2013.01.017}
}
\seealso{
\link{Lemna-models}
}
\keyword{datasets}
