% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{sequence}
\alias{sequence}
\title{Sequence of scenarios}
\usage{
sequence(seq)
}
\arguments{
\item{seq}{optional \code{vector} of \linkS4class{EffectScenario} objects}
}
\description{
A sequence of scenarios is treated as a single scenario and each scenario
is simulated one after the other. If scenario \code{n} in a sequence was simulated,
scenario \code{n+1} will start off in the state where \code{n} has ended.
Scenario sequences can be used to implement changes in e.g. model parameters
over time.
}
\details{
\subsection{Requirements}{

All scenarios in a sequence must fulfill the following requirements:
\itemize{
\item All scenarios must have identical state variables
\item The \emph{output times} of all scenarios must represent a continuous time series
without gaps or overlaps
}
}

\subsection{Limitations}{

Only simulation of sequences supported, yet. Calculation of effects or EPx
values not possible.
}
}
\examples{
# create two scenarios that need to be simulated one after the other
scen1 <- minnow_it \%>\% set_times(0:3)
scen2 <- minnow_it \%>\% set_times(3:6) \%>\% set_param(c(kd=0))

# create an empty sequence
sq <- sequence()
# assign scenarios to the sequence one by one
scenario(sq, 1) <- scen1
scenario(sq, 2) <- scen2

# simulate the sequence
simulate(sq)

# create a sequence and assign scenarios
sq <- sequence(list(scen1, scen2))
# assign list of scenarios
scenarios(sq) <- list(scen1, scen2)
}
