% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_batch}
\alias{simulate_batch}
\title{Returns data frame with simulation results}
\usage{
simulate_batch(model_base, treatments, param_sample = NULL)
}
\arguments{
\item{model_base}{effect scenario object with mean parameters}

\item{treatments}{treatments exposure levels as data frame (t, conc, trial)}

\item{param_sample}{data.frame with parameter sample}
}
\value{
a \code{data.frame}
}
\description{
Returns data frame with simulation results
}
